Board

Description

The file includes game chackerboard datatype, represented as Map from Board.Coorinates to Board.Field. It also includes functions to update the board and AI to apply moves automatically.

Documentation

invalidWolfPosition :: String

Error message, displayed if initial position of the wolf is not in the range 1..4.

newtype Board

Checkerboard datatype

Constructors

Board

Fields

board :: Map Coordinate Field

Map from Board.Coordinate to Board.Field

init

Arguments

:: Int
Initial wolf position. Must be positive integer from 1 to 4 and denotes wolf position on the first (upper) row of the board, from left to right.

-> Board
Initial board state.

Return a initial board to start the game.

moveStatus

Arguments

:: Board
Board state.

-> Move
Move to check.

-> Status
Move validity.

Checks if given move is valid on the board or is invalid for some reason.

(??) :: Board -> Move -> Status

Infix equivalent of moveStatus.

apply

Arguments

:: Board
Initial board state.

-> Move
Move to execute.

-> Board
New board state.

Execute a move on a board.

(>>>) :: Board -> Move -> Board

Infix equivalent of apply.

wolfCoord :: Board -> Coordinate

Return a wolf coordinates on the board.

sheepsCoords :: Board -> [Coordinate]

Return a list of sheep coordinates on the board.

validWolfMoves :: Board -> [Move]

List of moves of the wolf, that are valid in the current board state.

validSheepsMoves :: Board -> [Move]

List of moves of the sheeps, that are valid in the current board state.

wolfPoints :: Board -> Double

Calculates points, that wolf can use to rate its current situation in the game. With more points, the wolf situation is considered better. This function can be used as heuristics in mini-max algorithm and the game tree.

Changing the formula has great impact on the result of the game. Currently, wolf points are calculated as points = wolf distance from upper board edge + 2 * number of valid wolf moves + mean distance from wolf to sheeps

sheepsPoints :: Board -> Double

Points that sheeps can use to rate their current situation in the game. Calculated as -1 * wolfPoints.

bestMove

Arguments

:: (Board -> Double)
A function used to rate the board state. May be either wolfPoints or sheepsPoints.

-> Board
Current board state.

-> [Move]
List of moves to analyze. May be empty, if no moves are possible (one side won the game).

-> Maybe (Move, Double)
The best move, alogside with its rating. May be Nothing, if list of moves to analyze is empty.

From list of possible moves, return the best (the one that leads to the state of the board with biggest amount of points).

bestWolfMove

Arguments

:: Int
Depth of game tree to analyze.

-> Board
Current board state.

-> Maybe (Move, Double)
The best wolf move, alogside with its rating. May be Nothing, if list of moves to analyze is empty.

Auxiliary function, calculates the best wolf move on the current board state. Applies wolfPoints and validWolfMoves to bestMove.

TODO: add recursive analyze of game tree.

bestSheepMove

Arguments

:: Int
Depth of game tree to analyze.

-> Board
Current board state.

-> Maybe (Move, Double)
The best wolf move, alogside with its rating. May be Nothing, if list of moves to analyze is empty.

Auxiliary function, calculates the best sheeps move on the current board state. Applies sheepsPoints and validSheepsMoves to bestMove.

TODO: add recursive analyze of game tree

toFile

Arguments

:: Board
Board state to save.

-> String
File name.

-> IO ()

JSONize and save board to file

fromFile

Arguments

:: String
File name.

-> IO Board

Read board state from file