Board.Coordinate

Description

Datatype that contains valid coordinates of checkerboard fields.

Documentation

data Coordinate

Checkerboard has 8x8 size. Only black fields are used, so there are total 32 valid board fields and its coordinates. Also, out of the board coordinate is added, as moving a wolf or sheep may sometimes result in getting out of the board edges.

Constructors

A1

A3

A5

A7

B2

B4

B6

B8

C1

C3

C5

C7

D2

D4

D6

D8

E1

E3

E5

E7

F2

F4

F6

F8

G1

G3

G5

G7

H2

H4

H6

H8

OutOfBoard

move :: Coordinate -> Direction -> Coordinate

Calculate a new coordinate, after executing a move in specified direction.

(|>>) :: Coordinate -> Direction -> Coordinate

Infix equivalent of move.

distance :: Coordinate -> Coordinate -> Double

Calculate euclidean distance between two fields.