Board.Field

Description

Datatype that represents the board field. A field may be empty, or may be occupied by sheep or by a wolf.

Documentation

data Field

Datatype that represents the board field. A field may be empty, or may be occupied by sheep or by a wolf.

Constructors

Empty

Sheep

Wolf