Move.Direction

Description

Documentation

data Direction

On a checkerboard, a move can be done diagonally: up or down and left or right.

Constructors

DownLeft

DownRight

UpLeft

UpRight

data VAxis

On vertical axis, a move can be done up or down.

Constructors

Down

Up

data HAxis

On horizontal axis, a move can be done left or right.

Constructors

Left

Right

vaxis :: Direction -> VAxis

Return vertical direction of the move.

haxis :: Direction -> HAxis

Return horizontal direction of the move.